pyproject.toml
[tool.poetry]
name = "finance-rag-bench"
version = "0.1.0"
description = "Finance-focused RAG + evaluation benchmark with API, CLI, and agentic workflows."
authors = ["Finance RAG Team <finance@example.com>"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = { extras = ["standard"], version = "^0.29.0" }
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
python-dotenv = "^1.0.1"
structlog = "^24.1.0"
diskcache = "^5.6.3"
chromadb = "^0.4.24"
sentence-transformers = "^2.6.1"
openai = "^1.14.3"
anthropic = "^0.21.3"
google-generativeai = "^0.4.2"
azure-ai-openai = "^1.0.0"
rank-bm25 = "^0.2.2"
PyMuPDF = "^1.23.26"
python-docx = "^1.1.0"
pdfplumber = "^0.10.4"
pandas = "^2.2.1"
openpyxl = "^3.1.2"
numpy = "^1.26.4"
scipy = "^1.12.0"
typer = "^0.9.0"
rich = "^13.7.1"
tiktoken = "^0.6.0"
httpx = "^0.27.0"
regex = "^2023.12.25"
rouge-score = "^0.1.2"
scikit-learn = "^1.4.1"
jinja2 = "^3.1.3"
tenacity = "^8.2.3"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5"
pytest-cov = "^4.1.0"
anyio = "^4.3.0"
ruff = "^0.3.0"
types-requests = "^2.31.0.20240218"
mypy = "^1.8.0"

[tool.poetry.scripts]
ingest = "src.cli.ingest_cli:app"
ask = "src.cli.query_cli:app"
eval = "src.cli.eval_cli:app"

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
extend-select = ["I"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
disallow_untyped_defs = true
strict_optional = true
warn_unused_ignores = true
warn_return_any = true
